--- src/hooks/useCurrencyConverter/index.ts
@@ -1,1 +1,0 @@
-export { useCurrencyConverter } from "./useCurrencyConverter";
