+++ src/modules/ConversionTable/src/components/cells/AmountEURCell.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import { ConversionTableRow } from "../../models/ConversionTableRow";
 import { Box, Typography } from "@mui/joy";
 import { observer } from "mobx-react-lite";
-import { useCurrencyConverter } from "hooks/useCurrencyConverter";
+import { useCurrencyConverter } from "modules/CurrencyConversion";
 
 type AmountEURCellProps = { row: ConversionTableRow };
