+++ src/modules/CurrencyConversion/index.ts
@@ -1,0 +1,1 @@
+export { useCurrencyConverter } from "./src/hooks/useCurrencyConverter";
