+++ src/modules/CurrencyConversion/src/utils/convertEurToPln.ts
@@ -1,0 +1,4 @@
+import { cutDecimals } from "./utils/cutDecimals";
+
+export const convertEurToPln = (amountEur: number, eurToPlnRate: number) =>
+  cutDecimals(amountEur * eurToPlnRate);
