+++ src/modules/CurrencyConversion/src/utils/convertPlnToEur.ts
@@ -1,0 +1,4 @@
+import { cutDecimals } from "./utils/cutDecimals";
+
+export const convertPlnToEur = (amountPLN: number, eurToPlnRate: number) =>
+  cutDecimals(amountPLN / eurToPlnRate);
