+++ src/modules/CurrencyConversion/src/utils/utils/cutDecimals.ts
@@ -1,0 +1,2 @@
+export const cutDecimals = (overflowing: number) =>
+  parseFloat(overflowing.toFixed(2));
