--- src/utils/convertCurrency/index.ts
@@ -1,1 +1,0 @@
-export { convertEurToPln, convertPlnToEur } from "./convertCurrency";
